Fix printf() format specifier mismatch.
Use int for variable printed with "%d", not size_t.

+++ src/gui/wxComposeView.cpp
@@ -5861,7 +5861,7 @@ bool Composer::RestoreAll()
       return false;
    }
 
-   size_t nResumed = 0;
+   int nResumed = 0;
 
    wxString filename;
    bool cont = dir.GetFirst(&filename, wxEmptyString, wxDIR_FILES);
